separacao do builer e draw do BSPTree
